# Conflicts:
#	trunk/GoJava7/a.sukhenko/Kickstarter/src/main/java/ua/com/goit/gojava7/kickstarter/dao/db/QuoteDatabaseDao.java
#	trunk/GoJava7/a.sukhenko/Kickstarter/src/main/java/ua/com/goit/gojava7/kickstarter/domain/Category.java
#	trunk/GoJava7/a.sukhenko/Kickstarter/src/main/java/ua/com/goit/gojava7/kickstarter/domain/Project.java
#	trunk/GoJava7/a.sukhenko/Kickstarter/src/main/java/ua/com/goit/gojava7/kickstarter/domain/Question.java
#	trunk/GoJava7/a.sukhenko/Kickstarter/src/main/java/ua/com/goit/gojava7/kickstarter/domain/Quote.java

+++ trunk/GoJava7/a.smirnov/MyWebKickstarter/src/main/java/ua/com/goit/gojava7/kickstarter/dao/PaymentDao.java
@@ -8,5 +8,5 @@ public interface PaymentDao {
 
 	public void remove(Payment element);
 
-	public int getSumProjectPayments(int projectId);
+	public long getSumProjectPayments(int projectId);
 }
